feat(represent): a fidelity certificate names its referent, and the floors that read it are enforced
`RepresentationFloor::Exact` has ALWAYS meant "the deepest extent this
codec declares" — `admits` compared depths and never once consulted a
radius. For a lossless progressive codec the two readings coincide. For a
lossy one they are not the same claim: `VQ8_SHARED` at full depth is
every byte the artifact holds AND no bound whatever against the tensor it
was fitted to. This build would eventually have printed "exact
reconstruction" while planning terminal Q4 or VQ data.

THE REFERENT, settled and written down where the type is defined. A
certificate measures decoded values against the TYPED LOGICAL SOURCE
TENSOR presented at the representation boundary. Not against the stored
representation — that reading makes every terminal extent trivially
exact, since decoding all of an artifact always reproduces the artifact,
and collapses the distinction this commit exists to draw. Not including
kernel arithmetic — what a direct kernel over codes rounds to is a
property of the kernel and belongs to realization qualification.

The vocabulary follows, and the two questions are now asked separately:

  TerminalExtent  how much of the artifact is read. STRUCTURAL, states no
                  source-fidelity claim, still the default — so existing
                  plans keep their behaviour without inheriting a quality
                  claim nobody earned.
  CertifiedExact  a verified compatible certificate whose radius is 0.0.
  Within(r)       a verified composed certificate at or under r.

Neither evidence floor is satisfied by being terminal. `admits` no longer
short-circuits on depth for them: an operand with no certificate has made
no claim about its source, and no claim is not a small error.

DECLARATIONS BECOME HONEST rather than uniform. A lossless carrier can
say 0.0 and now does — raw `F32` is the identity, `F16`/`BF16` widen
exactly, `BF16_ZLIB` is byte-identical once inflated, and compression is
not approximation. A codec that FITS its source still says nothing:
`Q4_K`, `Q6_K`, `Q8_0`, `NVFP4`, `MXFP4` and `VQ8_SHARED` carry an error
that belongs to the instance encoded, and only an attestation can supply
it. The old contract test asserted "a single-extent codec declares no
radius", which conflated the two questions — it is not the extent count
that decides, it is whether anything was thrown away.

AND THE FLOORS ARE NOW ENFORCED, which is the half that made the rest
worth committing. `RepresentationFloor` was consulted in exactly one
place, `shallowest_saving`, which judges candidate MOVES under
preparation pressure. Under `ResidencyBudget::UNBOUNDED` there is no
pressure, so nothing consulted it at all: a caller could declare
`CertifiedExact`, plan a Q4_K model, and be refused nothing. The
requirement was real and no code ever asked it — the same hollow shape
step 6 removed from the composition itself.

Enforcement lands on the INITIAL selection, and deliberately nowhere
else. The only assignment to `extent.selected` after the initial pin
takes its extent from `shallowest_saving`, which already filters by
`fidelity.admits`, so reselection is enforced at the point of choice; a
re-check on the way out could never fire, and a refusal that cannot
happen is decoration. That reasoning is recorded at the return rather
than left for the next reader to re-derive.

The witnesses, on a real Q4_K container compiled from the fixture and
planned under a budget that asks for nothing:

  TerminalExtent  plans it, claiming nothing about it
  CertifiedExact  REFUSES, naming the operand, the representation, that
                  it states no source-fidelity bound, and that a verified
                  attestation is the one thing that could supply one
  Within(1e9)     REFUSES too — no bound is loose enough to buy a claim
                  that was never made

Plus the opposite half of the refusal: a composed bound that EXISTS and
is merely too large is refused by its number and metric, not by absence.
Removing the enforcement call turns both Q4 refusals red while the
structural arm stays green.

Structural, and confined to code this branch introduced: the step-6
carriage moves out of `prepared.rs` (already far over any sane length)
into `fidelity_carriage`, which now also owns enforcement — one module
for "what bound applies and who may rely on it", knowing nothing of
realizations, backends or budgets. `prepared.rs` drops 224 lines net. The
900-line `composed_floor.rs` becomes six files, none over 330, with every
premise in one `vocabulary` module so no value is spelled twice. No
pre-existing file was reorganised.

Gates: fmt, clippy (larql-vindex, larql-cli), cargo check against
x86_64-apple-darwin, cargo test -p larql-vindex (4878), cargo test -p
larql-cli (911), doc links. Coverage policy passed at 93.72%;
`fidelity_carriage.rs` came in at 89.77% and was brought over by testing
the branch it exposed and deleting one that could not be reached.
