+++ crates/larql-vindex/src/format/vindex3/opplan/exec/mod.rs
@@ -29,6 +29,7 @@ pub mod cpu;
 pub mod decode;
 pub mod device;
 mod experts;
+pub mod fidelity_carriage;
 pub mod gated_delta;
 pub mod hyper_connection;
 pub mod kda;
